add: use cases and fix app folders
