--- src/services/auth.service.ts
@@ -1,9 +1,0 @@
-import { api } from '@clients/index';
-import { UserModel } from '@models/user';
-
-const login = async (user: UserModel) => {
-  const { data } = await api.post<UserModel>('login', user);
-  return data;
-};
-
-export { login };
